CompleteBandalart, CompleteTopBar, SaveImageButton 컴포저블 함수 분리
