+++ feature/complete/src/main/kotlin/com/nexters/bandalart/android/feature/complete/CompleteScreen.kt
@@ -2,38 +2,20 @@ package com.nexters.bandalart.android.feature.complete
 
 import android.content.Intent
 import android.widget.Toast
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBackIos
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.airbnb.lottie.compose.LottieAnimation
@@ -43,14 +25,9 @@ import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
 import com.nexters.bandalart.android.core.ui.R
 import com.nexters.bandalart.android.core.ui.component.BandalartButton
-import com.nexters.bandalart.android.core.ui.component.EmojiText
-import com.nexters.bandalart.android.core.ui.component.FixedSizeText
 import com.nexters.bandalart.android.core.ui.component.TitleText
-import com.nexters.bandalart.android.core.ui.theme.Black
-import com.nexters.bandalart.android.core.ui.theme.Gray100
-import com.nexters.bandalart.android.core.ui.theme.Gray300
-import com.nexters.bandalart.android.core.ui.theme.Gray400
-import com.nexters.bandalart.android.core.ui.theme.Gray900
+import com.nexters.bandalart.android.feature.complete.ui.CompleteBandalart
+import com.nexters.bandalart.android.feature.complete.ui.CompleteTopBar
 
 @Composable
 internal fun CompleteRoute(
@@ -129,129 +106,20 @@ internal fun CompleteScreen(
         horizontalAlignment = Alignment.CenterHorizontally,
       ) {
         Spacer(modifier = Modifier.height(16.dp))
-        Row(
-          Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp),
-        ) {
-          IconButton(
-            onClick = onNavigateBack,
-            modifier = Modifier
-              .width(32.dp)
-              .aspectRatio(1f),
-          ) {
-            Icon(
-              imageVector = Icons.Default.ArrowBackIos,
-              contentDescription = context.getString(R.string.arrow_forward_descrption),
-              tint = Gray900,
-            )
-          }
-        }
+        CompleteTopBar(
+          context = context,
+          onNavigateBack = onNavigateBack,
+        )
         Spacer(modifier = Modifier.height(40.dp))
         TitleText(text = context.getString(R.string.complete_title))
         Box(modifier = Modifier.fillMaxSize()) {
-          Column(
+          CompleteBandalart(
             modifier = Modifier.align(Alignment.Center),
-            horizontalAlignment = Alignment.CenterHorizontally,
-          ) {
-            FixedSizeText(
-              text = context.getString(R.string.complete_chart),
-              color = Gray400,
-              fontWeight = FontWeight.W600,
-              fontSize = 14.sp,
-              letterSpacing = (-0.28).sp,
-            )
-            Spacer(modifier = Modifier.height(10.dp))
-            Box(
-              modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 33.dp)
-                .clip(RoundedCornerShape(12.dp))
-                .border(
-                  width = 2.dp,
-                  color = Gray300,
-                  shape = RoundedCornerShape(12.dp),
-                ),
-            ) {
-              Column(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalAlignment = Alignment.CenterHorizontally,
-              ) {
-                Spacer(modifier = Modifier.height(16.dp))
-                Card(
-                  shape = RoundedCornerShape(12.dp),
-                  elevation = CardDefaults.cardElevation(defaultElevation = 4.dp),
-                ) {
-                  Column {
-                    Box(
-                      modifier = Modifier
-                        .width(52.dp)
-                        .aspectRatio(1f)
-                        .background(Gray100),
-                      contentAlignment = Alignment.Center,
-                    ) {
-                      if (uiState.profileEmoji == context.getString(R.string.home_default_emoji)) {
-                        val image = painterResource(
-                          id = com.nexters.bandalart.android.core.designsystem.R.drawable.ic_empty_emoji,
-                        )
-                        Image(
-                          painter = image,
-                          contentDescription = context.getString(R.string.empty_emoji_descrption),
-                        )
-                      } else {
-                        EmojiText(
-                          emojiText = uiState.profileEmoji,
-                          fontSize = 22.sp,
-                        )
-                      }
-                    }
-                  }
-                }
-                Spacer(modifier = Modifier.height(6.dp))
-                FixedSizeText(
-                  text = uiState.title,
-                  color = Black,
-                  fontWeight = FontWeight.W700,
-                  fontSize = 16.sp,
-                  letterSpacing = (-0.32).sp,
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-              }
-            }
-          }
-//          // MVP 제외
-//          Box(
-//            modifier = Modifier
-//              .clickable(onClick = {})
-//              .align(Alignment.BottomCenter)
-//              .offset(y = (-110).dp),
-//          ) {
-//            Column {
-//              Row(verticalAlignment = Alignment.CenterVertically) {
-//                val image = painterResource(id = R.drawable.ic_gallery)
-//                Image(
-//                  painter = image,
-//                  contentDescription = "Option Icon",
-//                )
-//                Text(
-//                  text = buildAnnotatedString {
-//                    withStyle(
-//                      style = SpanStyle(
-//                        textDecoration = TextDecoration.Underline,
-//                      ),
-//                    ) {
-//                      append("이미지 저장하기")
-//                    }
-//                  },
-//                  color = Gray400,
-//                  fontFamily = pretendard,
-//                  fontWeight = FontWeight.W600,
-//                  fontSize = 16.sp.nonScaleSp,
-//                  letterSpacing = -(0.32).sp.nonScaleSp,
-//                )
-//              }
-//            }
-//          }
+            context = context,
+            uiState = uiState,
+          )
+          // MVP 제외
+          // SaveImageButton(modifier = Modifier.align(Alignment.BottomCenter))
           BandalartButton(
             onClick = shareBandalart,
             text = context.getString(R.string.complete_share),
